# ordertracker-angular/api.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
...
  "/api/save-product":
    post:
      tags:
      - product-controller
      operationId: saveProduct
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Product"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Product"
...
  "/api/add-product":
    post:
      tags:
      - product-controller
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Product"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Product"
...
  "/api/find-product-by-id":
    get:
      tags:
      - product-controller
      operationId: findProductById
      parameters:
      - name: findProductByIdDto
        in: query
        required: true
        schema:
          "$ref": "#/components/schemas/FindProductByIdDto"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/Product"
components:
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - GENERAL
          - THUMBNAIL
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: integer
          format: int32
        images:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/Image"
...
    FindProductByIdDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        withBinary:
          type: boolean