# ordertracker-angular/src/app/api/ordertracker-backend/model/Image.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Image {
    id?: number;
    type?: Image.TypeEnum;
    imageData?: string;
}
export namespace Image {
    export type TypeEnum = 'GENERAL' | 'THUMBNAIL';
    export const TypeEnum = {
        GENERAL: 'GENERAL' as TypeEnum,
        THUMBNAIL: 'THUMBNAIL' as TypeEnum
    };
}